three new cogs log welcome youtube some changes in main

+++ data/db/build.sql
@@ -1,5 +1,5 @@
 CREATE TABLE IF NOT EXISTS exp (
-	UserID integer PRIMARY KEY,
+	UserID  bigint PRIMARY KEY,
 	XP integer DEFAULT 0,
 	Level integer DEFAULT 0,
 	XPLock text DEFAULT CURRENT_TIMESTAMP
